Initial commit. And some thoughts...
Shrinkwrap is the keystone of this package. It allows to dig inside
`@opma-alpha/tgls` and force it to depend on the same version of `tsdl`
than us, instead of having to fork `tgls` and manually update it.
Special thanks to @jordwalke for coming up with the idea.

This exposes opengl and webgl, which should follow the exact same
interface describe in src/gl.re. You only need to have a functor that
takes a module with the GL signature and the world is yours.

This will allow any package to use GL capabilities on both native and
web without any special casing code.

TODO: there's one place where we have special casing, which is in the
versioning of the shaders. This will be fixed soon!

@schmavery @chenglou :D