Huge change. Support textures + bunch of small functions + added C dep.
This was made and tested inside `reprocessing` first.
I added a couple GL functions related to textures, and a C dependency
called SOIL to allow people to load images from disk. That library
supports a whole bunch of formats:
https://github.com/kbranigan/Simple-OpenGL-Image-Library.

We could clean up the SOIL code as we might not need all of it. I just
downloaded it and kept the folder structure.

One thing to note is `loadImage` in native allows you to _force_ pixel
color formats (RGB, RGBA etc...) but in web doesn't. In web we use
`new Image()` and set the `src` field. Seeing what people do online, it
seems like you can assume RGBA and the system will handle the edge
cases. I decided to leave `loadOption` as a parameter to `loadImage` so
that someone who would only use the native part could have that control.
This opens a new door to all the things that don't fall into _both_
native and web, and what we should do about them. I'd like to avoid
removing control because one platform doesn't support it. I prefer
instead to leave extra arguments that are optional for one platform.